+++ device/containers.go
@@ -12,6 +12,7 @@ import (
 	"github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/client"
 	"github.com/docker/go-connections/nat"
+	"github.com/shamanec/GADS-devices-provider/models"
 	"github.com/shamanec/GADS-devices-provider/util"
 	log "github.com/sirupsen/logrus"
 )
@@ -438,4 +439,24 @@ func (device *LocalDevice) createAndroidContainer() {
 	}
 	// Delete the container from the map with containers being created
 	delete(createdContainers, device.Device.UDID)
+}
+
+// Check if device has an existing container
+func (device *LocalDevice) hasContainer(allContainers []types.Container) (bool, error) {
+	for _, container := range allContainers {
+		// Parse plain container name
+		containerName := strings.Replace(container.Names[0], "/", "", -1)
+
+		if strings.Contains(containerName, device.Device.UDID) {
+			deviceContainer := models.DeviceContainer{
+				ContainerID:     container.ID,
+				ContainerStatus: container.Status,
+				ImageName:       container.Image,
+				ContainerName:   containerName,
+			}
+			device.Device.Container = &deviceContainer
+			return true, nil
+		}
+	}
+	return false, nil
 }
